% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{attr_on}
\alias{attr_on}
\title{Attributes On}
\usage{
attr_on(...)
}
\arguments{
\item{...}{characters indicating attributes to turn on. \code{"bf"} for bold face; \code{"ft"} for faint; \code{"it"} for italics; \code{"ul"} for underline; \code{"sb"} for slow blink; \code{"fb"} for fast blink; \code{"rv"} for reverse video (invert bg and fg colors); \code{"st"} for strike-through}
}
\value{
\code{NULL}
}
\description{
Turns on text attributes in terminal, including bold text, italics, underline, etc. Note that not all terminals support each attribute.
}
\details{
Use \code{\link{attr_off}} to turn off the attributes.
}
\examples{
cat("hello world!\n")
attr_on("bf", "ul")
cat("hello world!\n")
attr_off()

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
