% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{bg_off}
\alias{bg_off}
\title{Turn Off Background Color}
\usage{
bg_off()
}
\value{
\code{NULL}
}
\description{
Return the background of future terminal text to the default color. Background color is turned on with \code{\link{bg_on}}.
}
\examples{
# Different methods of specifying yellow
bg_on("yellow")
bg_on("#FFFF00")
bg_on(11)
bg_on(255, 255, 0)

# Turn off color
bg_off()

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
