% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{clear}
\alias{clear}
\title{Clear Text}
\usage{
clear(x = c("screen", "end", "start"), ...)
}
\arguments{
\item{x}{character describing console location to clear. The default, \code{"screen"}, clears the entire screen; \code{"start"} clears all text from the beginning of the screen until the cursor's position; \code{"end"} clears all text from the cursor's position to the bottom of the screen.}

\item{...}{objects passed to/from methods}
}
\value{
\code{NULL}
}
\description{
Clear text from the terminal. Passing values \code{"start"} or \code{"end"} allow the user to clear specific portions of the screen relative to the cursor.
}
\examples{
clear()

cat(paste(LETTERS[1:10], collapse="\n"))

clear("start")
clear("end")

}
