% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{color_off}
\alias{color_off}
\title{Turn Off Colors in Terminal}
\usage{
color_off()
}
\value{
\code{NULL}
}
\description{
Return the background and foreground of future terminal text to the default colors.
}
\examples{
bg_on("red")
fg_on("yellow")

# Turn off color
color_off()

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
