% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{erase}
\alias{erase}
\title{Erase Text}
\usage{
erase(x = c("row", "start", "end"), ...)
}
\arguments{
\item{x}{character describing location to clear. The default, \code{"row"}, clears the entire row; \code{"start"} clears all text from the beginning of the row until the cursor's position; \code{"end"} clears all text from the cursor's position until the end of the row.}

\item{...}{objects passed to/from methods}
}
\value{
\code{NULL}
}
\description{
Clear text from the cursor's row . Passing values \code{"start"} and \code{"end"} allow the user to erase specific portions of the row relative to the cursor.
}
\examples{
cat('hello world!')
erase('row')

}
