% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-Example.R
\name{example_luckynumber}
\alias{example_luckynumber}
\title{Example Program From Vignette}
\usage{
example_luckynumber()
}
\value{
\code{NULL}
}
\description{
Simple program that asks for a letter and a number and returns another value to screen.
}
