% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{fg_off}
\alias{fg_off}
\title{Turn Off Foreground Color}
\usage{
fg_off()
}
\value{
\code{NULL}
}
\description{
Return future terminal text to the default color. Foreground color is turned on with \code{\link{fg_on}}.
}
\examples{
# Different methods of specifying red
fg_on("red")
fg_on("#FF0000")
fg_on(1)
fg_on(255, 0, 0)

# Turn off color
fg_off()

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
