% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{fg_on}
\alias{fg_on}
\title{Turn On Foreground Color}
\usage{
fg_on(...)
}
\arguments{
\item{...}{character or numeric value}
}
\value{
\code{NULL}
}
\description{
Specifies the color of all future text written in the terminal \code{fg_on} accepts numeric values (RGB or 8-bit color code), hexadecimal characters, or the name of the color. Not all terminals support each possible color.
}
\details{
Foreground color is turned off with \code{\link{fg_off}}.
}
\examples{
# Different methods of specifying red
fg_on("red")
fg_on("#FF0000")
fg_on(9)
fg_on(255, 0, 0)

# Turn off color
fg_off()

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
