% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-Draw.R
\name{fill_rect}
\alias{fill_rect}
\title{Draw a Filled-In Rectangle}
\usage{
fill_rect(yx1, yx2, text = "x", ...)
}
\arguments{
\item{yx1}{upper-left \code{(row, col)} coordinate}

\item{yx2}{lower-right \code{(row, col)} coordinate}

\item{text}{character value drawn at coordinate}

\item{...}{parameters that are passed to \code{style()}, including the foreground color \code{fg}, background color \code{bg}, and attribute \code{attr}}
}
\value{
\code{NULL}
}
\description{
Calculate the path of a rectangle in a grid and draw to screen.
}
\examples{
draw_rect(c(5,5), c(9,9))

}
\seealso{
Other drawing functions: 
\code{\link{box_at}()},
\code{\link{draw_arc}()},
\code{\link{draw_bezier}()},
\code{\link{draw_circle}()},
\code{\link{draw_ellipse}()},
\code{\link{draw_fn}()},
\code{\link{draw_lerp}()},
\code{\link{draw_path}()},
\code{\link{draw_ray}()},
\code{\link{draw_rect}()},
\code{\link{draw_shape}()},
\code{\link{fill_circle}()},
\code{\link{fill_ellipse}()},
\code{\link{fill_shape}()},
\code{\link{grid_at}()},
\code{\link{grid_mat}()},
\code{\link{hline_at}()},
\code{\link{hline}()},
\code{\link{vline_at}()},
\code{\link{vline}()}
}
\concept{drawing functions}
