% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{hide_cursor}
\alias{hide_cursor}
\title{Hide Cursor}
\usage{
hide_cursor()
}
\value{
\code{NULL}
}
\description{
Make the cursor invisible. The cursor can be revealed with \code{\link{show_cursor}}
}
\examples{
hide_cursor()
cat("\n\nHello World!")
show_cursor()

}
