% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{in.term}
\alias{in.term}
\title{Determine whether in Terminal}
\usage{
in.term()
}
\value{
logical value; \code{TRUE} or \code{FALSE}
}
\description{
Tests whether the session is in terminal and returns TRUE or FALSE. Many of the \code{cursr} functions require being in terminal and will not work with RStudio or the R GUI application.
}
\examples{
in.term()

}
