% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{load_cursor}
\alias{load_cursor}
\title{Load Cursor}
\usage{
load_cursor()
}
\value{
\code{NULL}
}
\description{
Restore cursor to its previously saved location from \code{\link{save_cursor}}.
}
\examples{
save_cursor()
cat("\n\nHello World!")
load_cursor()

}
