% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{make_bg}
\alias{make_bg}
\title{Create Background Color}
\usage{
make_bg(...)
}
\arguments{
\item{...}{character or numeric value}
}
\value{
ANSI character string
}
\description{
Returns the ANSI code for the specified background color. \code{make_bg} accepts numeric values (RGB or 8-bit color code), hexadecimal characters, or the name of the color.
}
\examples{
# Different methods of specifying cyan
make_bg("cyan")
make_bg("#00FFFF")
make_bg(14)
make_bg(0, 255, 255)

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
