% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{make_fg}
\alias{make_fg}
\title{Create Foreground Color}
\usage{
make_fg(...)
}
\arguments{
\item{...}{character or numeric value}
}
\value{
ANSI character string
}
\description{
Returns the ANSI code for the specified foreground color. \code{make_fg} accepts numeric values (RGB or 8-bit color code), hexadecimal characters, or the name of the color.
}
\examples{
# Different methods of specifying red
make_fg("red")
make_fg("#FF0000")
make_fg(9)
make_fg(255, 0, 0)

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_style}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
