% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{make_style}
\alias{make_style}
\title{Create Color & Attribute Style}
\usage{
make_style(fg = NA, bg = NA, attr = NA)
}
\arguments{
\item{fg}{character or numeric value for the foreground color. See \code{\link{fg_on}} for more details.}

\item{bg}{character or numeric value for the background color. See \code{\link{bg_on}} for more details.}

\item{attr}{character vector describing attributes to turn on. See \code{\link{attr_on}} for more details.}
}
\value{
ANSI character string
}
\description{
Returns the ANSI codes for the specified colors and text attributes.
}
\examples{
cat(make_style(fg="blue", bg=c(192,192,192), attr=c("ul", "st")))
cat("Hello World!\n")
reset()

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{reset}()},
\code{\link{style}()}
}
\concept{style functions}
