% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-Move.R
\name{mv_col}
\alias{mv_col}
\title{Move Cursor to Column}
\usage{
mv_col(n = 1L)
}
\arguments{
\item{n}{positive integer specifying the column}
}
\value{
\code{NULL}
}
\description{
Move the cursor to the specified column, while maintaining the same row.
}
\details{
The user must be in a terminal to use the functionality; it will not work in RStudio or the R GUI.
}
\examples{
# Position cursor at the beginning of the row
mv_col(1)

# Move cursor to the 10th column in the row
mv_col(10)

}
\seealso{
Other moving functions: 
\code{\link{mv_row}()},
\code{\link{mv_to}()},
\code{\link{mv}()}
}
\concept{moving functions}
