% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Path.R
\name{path_circle}
\alias{path_circle}
\title{Path of a Circle}
\usage{
path_circle(yx, r = 1, n = 50)
}
\arguments{
\item{yx}{center \code{(row, col)} coordinate}

\item{r}{radius of the circle in grid points}

\item{n}{number of points along curve to calculate}
}
\value{
Nx2 matrix of \code{(row, column)} coordinates
}
\description{
Calculate the path of a circle in a grid.
}
\examples{
path_circle(yx=c(10,10), r=5)

}
\seealso{
Other path-fitting functions: 
\code{\link{path_arc}()},
\code{\link{path_bezier}()},
\code{\link{path_ellipse}()},
\code{\link{path_fill}()},
\code{\link{path_fn}()},
\code{\link{path_intersection}()},
\code{\link{path_lerp}()},
\code{\link{path_ray}()},
\code{\link{path_rect}()},
\code{\link{path_shape}()}
}
\concept{path-fitting functions}
