% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Path.R
\name{path_fill}
\alias{path_fill}
\title{Fill In Path}
\usage{
path_fill(mat)
}
\arguments{
\item{mat}{Nx2 matrix of \code{(row, column)} path coordinates}
}
\value{
Nx2 matrix of \code{(row, column)} coordinates
}
\description{
Calculate the coordinates of all points inside of a path.
}
\examples{
c0 <- path_circle(c(10,10), r=5)
path_fill(c0)

}
\seealso{
Other path-fitting functions: 
\code{\link{path_arc}()},
\code{\link{path_bezier}()},
\code{\link{path_circle}()},
\code{\link{path_ellipse}()},
\code{\link{path_fn}()},
\code{\link{path_intersection}()},
\code{\link{path_lerp}()},
\code{\link{path_ray}()},
\code{\link{path_rect}()},
\code{\link{path_shape}()}
}
\concept{path-fitting functions}
