% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Path.R
\name{path_intersection}
\alias{path_intersection}
\title{Intersection between Two Paths}
\usage{
path_intersection(path)
}
\arguments{
\item{path}{list containing two coordinate \code{(row, column)} matrices.}
}
\value{
Nx2 matrix of \code{(row, column)} coordinates
}
\description{
Calculate the points of intersection between two paths.
}
\examples{
c1 <- path_circle(c(4,4), r=3)
c2 <- path_circle(c(6,6), r=3)
path_intersection(list(c1, c2))

}
\seealso{
Other path-fitting functions: 
\code{\link{path_arc}()},
\code{\link{path_bezier}()},
\code{\link{path_circle}()},
\code{\link{path_ellipse}()},
\code{\link{path_fill}()},
\code{\link{path_fn}()},
\code{\link{path_lerp}()},
\code{\link{path_ray}()},
\code{\link{path_rect}()},
\code{\link{path_shape}()}
}
\concept{path-fitting functions}
