% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Path.R
\name{path_lerp}
\alias{path_lerp}
\title{Linear Interpolation Path}
\usage{
path_lerp(start, end, n = 50)
}
\arguments{
\item{start}{starting \code{(row, col)} coordinate}

\item{end}{ending \code{(row, col)} coordinate}

\item{n}{number of points along curve to calculate}
}
\value{
Nx2 matrix of \code{(row, column)} coordinates
}
\description{
Interpolate between two points in a grid.
}
\examples{
path_lerp(start=c(10,1), end=c(1,3))

}
\seealso{
Other path-fitting functions: 
\code{\link{path_arc}()},
\code{\link{path_bezier}()},
\code{\link{path_circle}()},
\code{\link{path_ellipse}()},
\code{\link{path_fill}()},
\code{\link{path_fn}()},
\code{\link{path_intersection}()},
\code{\link{path_ray}()},
\code{\link{path_rect}()},
\code{\link{path_shape}()}
}
\concept{path-fitting functions}
