% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Path.R
\name{path_rect}
\alias{path_rect}
\title{Rectangle Path}
\usage{
path_rect(yx1, yx2)
}
\arguments{
\item{yx1}{upper-left \code{(row, col)} coordinate}

\item{yx2}{lower-right \code{(row, col)} coordinate}
}
\value{
Nx2 matrix of \code{(row, column)} coordinates
}
\description{
Calculate the path of a rectangle in a grid.
}
\examples{
path_rect(c(5,5), c(9,9))

}
\seealso{
Other path-fitting functions: 
\code{\link{path_arc}()},
\code{\link{path_bezier}()},
\code{\link{path_circle}()},
\code{\link{path_ellipse}()},
\code{\link{path_fill}()},
\code{\link{path_fn}()},
\code{\link{path_intersection}()},
\code{\link{path_lerp}()},
\code{\link{path_ray}()},
\code{\link{path_shape}()}
}
\concept{path-fitting functions}
