% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Write.R
\name{repch}
\alias{repch}
\title{Repeat a Character}
\usage{
repch(x, n)
}
\arguments{
\item{x}{character to be repeated}

\item{n}{number of times to be repeated}
}
\value{
character vector
}
\description{
Repeat a character \code{n} times and concatenate into a single value.
}
\examples{
repch("abc", 5)

}
