% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Style.R
\name{reset}
\alias{reset}
\title{Reset Console Style}
\usage{
reset()
}
\value{
\code{NULL}
}
\description{
Turns off all text attributes and colors in the terminal.
}
\examples{
attr_on("ul")
fg_on("red")
bg_on(c(10,60,205))
cat("Hello World!\n")

reset()
cat("Hello World!\n")

}
\seealso{
Other style functions: 
\code{\link{attr_off}()},
\code{\link{attr_on}()},
\code{\link{bg_off}()},
\code{\link{bg_on}()},
\code{\link{color_off}()},
\code{\link{color_pair}()},
\code{\link{fg_off}()},
\code{\link{fg_on}()},
\code{\link{make_bg}()},
\code{\link{make_fg}()},
\code{\link{make_style}()},
\code{\link{style}()}
}
\concept{style functions}
