% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{save_cursor}
\alias{save_cursor}
\title{Save Cursor Position}
\usage{
save_cursor()
}
\value{
\code{NULL}
}
\description{
Save the position of the cursor so that the position can be restored for later with \code{\link{load_cursor}}.
}
\examples{
save_cursor()
cat("\n\nHello World!")
load_cursor()

}
