% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{show_cursor}
\alias{show_cursor}
\title{Show Cursor}
\usage{
show_cursor()
}
\value{
\code{NULL}
}
\description{
Reveal the cursor after it has been hidden by \code{\link{hide_cursor}}.
}
\examples{
hide_cursor()
cat("\n\nHello World!")
show_cursor()

}
