% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-Sym.R
\docType{data}
\name{Sym}
\alias{Sym}
\title{Unicode Symbols}
\format{A named list of characters}
\usage{
Sym
}
\description{
A named list containing the unicode character for various box drawing, mathematical, currency, astrological, and other symbols.
}
\keyword{datasets}
