% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{term_dim}
\alias{term_dim}
\title{Determine Terminal Size}
\usage{
term_dim()
}
\value{
numeric vector (# of rows, # of columns)
}
\description{
Function determines the size of the terminal in number of rows and columns. The value may not be accurate in RStudio or the R GUI.
}
\examples{
term_dim()

}
