% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-Draw.R
\name{vline}
\alias{vline}
\title{Vertical Line}
\usage{
vline(n, text = "|")
}
\arguments{
\item{n}{integer describing the character length of the line}

\item{text}{character to be repeated}
}
\value{
character string of length \code{n}, separated by \code{"\n"}
}
\description{
Vertical Line
}
\examples{
vline(4, "*") # "*\n*\n*\n*"

}
\seealso{
Other drawing functions: 
\code{\link{box_at}()},
\code{\link{draw_arc}()},
\code{\link{draw_bezier}()},
\code{\link{draw_circle}()},
\code{\link{draw_ellipse}()},
\code{\link{draw_fn}()},
\code{\link{draw_lerp}()},
\code{\link{draw_path}()},
\code{\link{draw_ray}()},
\code{\link{draw_rect}()},
\code{\link{draw_shape}()},
\code{\link{fill_circle}()},
\code{\link{fill_ellipse}()},
\code{\link{fill_rect}()},
\code{\link{fill_shape}()},
\code{\link{grid_at}()},
\code{\link{grid_mat}()},
\code{\link{hline_at}()},
\code{\link{hline}()},
\code{\link{vline_at}()}
}
\concept{drawing functions}
