% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Write.R
\name{wr}
\alias{wr}
\title{Write String to Terminal}
\usage{
wr(text, fg = NA, bg = NA, attr = NA)
}
\arguments{
\item{text}{string to be printed to the Console}

\item{fg}{foreground color. See \code{\link{fg_on}} for more details.}

\item{bg}{background color. See \code{\link{bg_on}} for more details.}

\item{attr}{character attribute. See \code{\link{attr_on}} for more details.}
}
\value{
\code{NULL}
}
\description{
Writes a string of characters to the terminal at the current cursor position. \code{wr} accepts text colors and attributes, but these are reset to default afterwards if used.
}
\examples{
mv_to(5,4)
wrch("h")
wrch("e", fg="red")
wr("llo World")

}
\seealso{
Other writing functions: 
\code{\link{wrat}()},
\code{\link{wrch}()},
\code{\link{wrkpl}()},
\code{\link{wrkp}()}
}
\concept{writing functions}
