% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-BasicTerminal.R
\name{wrapup}
\alias{wrapup}
\title{Return Screen to Blank State}
\usage{
wrapup()
}
\value{
\code{NULL}
}
\description{
Function to be used at the end of a terminal function. It resets the colors and attributes to their default values, clears the screen, and reveals the cursor.
}
