% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Write.R
\name{wrchat}
\alias{wrchat}
\title{Write Character to Terminal at Specified Location}
\usage{
wrchat(row, col, chr, ...)
}
\arguments{
\item{row}{row in which character is printed. If length of \code{row} is greater than one, the second value replaces \code{col}.}

\item{col}{column in which character is printed}

\item{chr}{character to be printed to the Console}

\item{...}{parameters that are passed to \code{style()}, including the foreground color \code{fg}, background color \code{bg}, and attribute \code{attr}}
}
\value{
\code{NULL}
}
\description{
Move cursor to specified location in the terminal screen, then print the supplied character. This function will only work in terminal, not the RStudio Console or R GUI.
}
\details{
The coordinates are given in matrix notation: \code{(row, column)}, with the top-left corner of the screen being \code{(1,1)}.
}
\examples{
wrchat(5, 4, "h")

}
